Wrap context-aware handlers + values around loggers by default. (#31)
Previously, we didn't allowed you to provide any handler you wanted,
which didn't quite meet expectations for using clog. This changes the
behavior to always wrap the logger with a context handler.

Additionally, when a new logger is created any values present in the
context are added to the logger, which should allow the values to be
preserved even if the default slog funcs are used.