gcp: Surface trace propagation functions on context (#40)
Users of clog have to be able to reliably put trace information into the
context to make the logger pick it up correctly. The current
implementation uses a string, which is prone to collisions.

This surfaces context manipulation functions instead that operate on a
private type and allow for easier integration.